+++ R/goldrake-package.R
@@ -1,3 +1,7 @@
 #' @keywords internal
-#' @return a [tibble][tibble::tibble-package]
+#' @aliases "goldrake-package"
+#' @import fs
+#' @importFrom glue glue glue_collapse
+#' @importFrom here here
+#' @importFrom tibble tibble
 "_PACKAGE"