+++ NEWS.md
@@ -1,6 +1,16 @@
+* Added `add_reviewer()`
+* Added `print()` method for goldrake
+* Added `balanced_group_by()`
+* Added `set_gold_classes()`
+* Added `goldrake()` constructor
+* Added support for tidy evaluation
+* Updated `README`;
+* Added `utils.R`, `directory.R`, and `ui.R` with some feature from 
+  the package `usethis`.
+
 # goldrake 0.0.0.9000
 
-* Update Dependencies (R >= 3.4)
+* Update Dependencies (R >= 3.4);
 * Added basic development support:
   - git + GitHub;
   - appVeyor + Travis + codecov;